+++ frontend/package.json
@@ -18,6 +18,7 @@
     "antd-img-crop": "^4.12.2",
     "axios": "^1.4.0",
     "bootstrap": "^5.2.3",
+    "css-loader": "^6.8.1",
     "daisyui": "^2.42.1",
     "google-maps-react": "^2.0.6",
     "react": "^18.2.0",
