+++ frontend/src/container/ForgotPassword/ForgotPassword.js
@@ -1,10 +1,8 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import { Form, Input, Button, Row, Col, Typography } from "antd";
 import { useNavigate } from "react-router-dom";
-import axios from "../api";
-import "./Register.css";
-import lostFound from "../Pic/lost&found.png";
+import lostFound from "../../Pic/lost&found.png";
 const Container = styled(Row)(() => ({
   width: "100vw",
   height: "100vh",
@@ -29,35 +27,8 @@ const LoginButton = styled(Button)(() => ({
   alignItems: "center",
   gap: "1vmin",
 }));
-const Register = () => {
+const ForgotPassword = () => {
   const navigate = useNavigate();
-  const [id, setId] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [checkPassword, setCheckPassword] = useState("");
-
-  const register = async () => {
-    if (!/^[a-zA-z][0-9]{8}/.test(id)) {
-      console.log("invalid student id");
-    } else if (
-      !/^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/.test(
-        email
-      )
-    ) {
-      console.log("invalid email");
-    } else if (password !== checkPassword) {
-      console.log("wrong checkpassword");
-    } else {
-      const {
-        data: { result },
-      } = await axios.post("/register", {
-        id,
-        email,
-        password,
-      });
-      console.log(result);
-    }
-  };
   return (
     <Container>
       <Logo className="left" span={12}>
@@ -88,14 +59,8 @@ const Register = () => {
       <Col span={12}>
         <Row justify="center" align="middle" style={{ height: "100vh" }}>
           <Col span={14}>
-            <h1 className="fontContainer_2">Newly Arrived! Register Here</h1>
+            <h1 className="fontContainer_2">Oops&thinsp;‼️ Forgot Password</h1>
             <Form>
-              <Form.Item className="fontContainer_2">
-                <Typography style={{ textAlign: "left" }}>
-                  &thinsp;Student ID
-                </Typography>
-                <Input value={id} onChange={(e) => setId(e.target.value)} />
-              </Form.Item>
               <Form.Item>
                 <Typography
                   className="fontContainer_2"
@@ -103,31 +68,7 @@ const Register = () => {
                 >
                   &thinsp;Email
                 </Typography>
-                <Input
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                />
-              </Form.Item>
-              <Form.Item>
-                <Typography
-                  className="fontContainer_2"
-                  style={{ textAlign: "left" }}
-                >
-                  &thinsp;Password
-                </Typography>
-                <Input.Password
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                />
-              </Form.Item>
-              <Form.Item className="fontContainer_2">
-                <Typography style={{ textAlign: "left" }}>
-                  &thinsp;Check Password
-                </Typography>
-                <Input.Password
-                  value={checkPassword}
-                  onChange={(e) => setCheckPassword(e.target.value)}
-                />
+                <Input />
               </Form.Item>
               <Form.Item>
                 <div
@@ -140,18 +81,18 @@ const Register = () => {
                   }}
                 >
                   <LoginButton
-                    className="btn"
+                    className="fontContainer_2 btn"
                     block
                     onClick={() => navigate("/")}
                   >
-                    <Typography className="back_btn">←</Typography>
+                    <Typography>Cancel</Typography>
                   </LoginButton>
                   <LoginButton
                     className="fontContainer_2 btn"
                     block
-                    onClick={() => register()}
+                    onClick={() => navigate("/")}
                   >
-                    <Typography>Register</Typography>
+                    <Typography>Send</Typography>
                   </LoginButton>
                 </div>
               </Form.Item>
@@ -163,4 +104,4 @@ const Register = () => {
   );
 };
 
-export default Register;
+export default ForgotPassword;
