+++ frontend/src/container/Login/Login.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import { Form, Input, Button, Row, Col, Divider, Typography } from "antd";
 import { useNavigate } from "react-router-dom";
-import googleIcon from "../Pic/google-logo.png";
-import lostFound from "../Pic/lost&found.png";
+import googleIcon from "../../Pic/google-logo.png";
+import lostFound from "../../Pic/lost&found.png";
 import { useGoogleLogin } from "@react-oauth/google";
 import axios from "axios";
 import "./Login.css";
