+++ frontend/src/container/Uploadpage/Uploadpage.js
@@ -1,9 +1,9 @@
-import { InfoForm } from "../components/Form";
-import UpMap from "../components/UpMap";
+import { InfoForm } from "../../components/Form";
+import UpMap from "../../components/UpMap";
 import React, { useState } from "react";
 import { Col, Row } from "antd";
 import { useLoadScript } from "@react-google-maps/api";
-import Sidebar from "../components/Sidebar";
+import Sidebar from "../../components/Sidebar";
 const Upload = () => {
   const { isLoaded } = useLoadScript({
     googleMapsApiKey: "AIzaSyAaZZfGnw5Aud0RxgRgc3-G-db_7z-tptk", // Add your API key//AIzaSyAaZZfGnw5Aud0RxgRgc3-G-db_7z-tptk
